add: add margin and padding to `Featured` and create item for each (#12)
